backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
openai
anthropic
groq
ruff
black
python-multipart
aiopath
